Extract FileIconData.kt
