+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/util/MimeTypeIconProvider.kt
@@ -16,7 +16,7 @@
 
 package io.getstream.chat.android.compose.ui.util
 
-import io.getstream.chat.android.compose.R
+import io.getstream.chat.android.compose.ui.components.attachments.files.FileIconData
 import io.getstream.chat.android.models.AttachmentType
 import io.getstream.chat.android.ui.common.model.MimeType
 
@@ -74,32 +74,4 @@ internal object MimeTypeIconProvider {
             else -> FileIconData.Generic
         }
     }
-}
-
-internal enum class FileIconData(val resId: Int, val typeName: String) {
-    Pdf(R.drawable.stream_compose_ic_file_pdf, "pdf"),
-    Csv(R.drawable.stream_compose_ic_file_spreadsheet, "csv"),
-    Xls(R.drawable.stream_compose_ic_file_spreadsheet, "xls"),
-    Xlsx(R.drawable.stream_compose_ic_file_spreadsheet, "xlsx"),
-    Tar(R.drawable.stream_compose_ic_file_compression, "tar"),
-    Zip(R.drawable.stream_compose_ic_file_compression, "zip"),
-    Rar(R.drawable.stream_compose_ic_file_compression, "rar"),
-    G7z(R.drawable.stream_compose_ic_file_compression, "7z"),
-    Doc(R.drawable.stream_compose_ic_file_doc, "doc"),
-    Docx(R.drawable.stream_compose_ic_file_doc, "docx"),
-    Txt(R.drawable.stream_compose_ic_file_doc, "txt"),
-    Rtf(R.drawable.stream_compose_ic_file_doc, "rtf"),
-    Odt(R.drawable.stream_compose_ic_file_doc, "odt"),
-    Ppt(R.drawable.stream_compose_ic_file_presentation, "ppt"),
-    Pptx(R.drawable.stream_compose_ic_file_presentation, "pptx"),
-    Html(R.drawable.stream_compose_ic_file_code, "html"),
-    Md(R.drawable.stream_compose_ic_file_code, "md"),
-    M4a(R.drawable.stream_compose_ic_file_audio, "m4a"),
-    Mp3(R.drawable.stream_compose_ic_file_audio, "mp3"),
-    Aac(R.drawable.stream_compose_ic_file_audio, "aac"),
-    Mov(R.drawable.stream_compose_ic_file_video, "mov"),
-    Mp4(R.drawable.stream_compose_ic_file_video, "mp4"),
-    Generic(R.drawable.stream_compose_ic_file_generic, ""),
-    AudioGeneric(R.drawable.stream_compose_ic_file_audio, "audio"),
-    VideoGeneric(R.drawable.stream_compose_ic_file_video, "video"),
 }
